pyproject.toml
[project]
name = "my_project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "fastapi==0.104.1",
    "ollama>=0.3.3",
    "openpyxl==3.1.2",
    "pandas==2.1.3",
    "pydantic==2.5.0",
    "pypdf2==3.0.1",
    "python-docx==1.1.0",
    "python-multipart==0.0.6",
    "python-pptx==0.6.23",
    "qdrant-client >=1.13.0",
    "uvicorn==0.24.0",
    "pydantic-settings<=2.5.0",
]
